Extract ToHbmFilterParam as an extension
