+++ src/FluentNHibernate/Utils/HbmExtensions.cs
@@ -1,15 +1,26 @@
-﻿using FluentNHibernate.MappingModel;
+﻿using System.Collections.Generic;
+using FluentNHibernate.MappingModel;
 using NHibernate.Cfg.MappingSchema;
+using NHibernate.Type;
 
 namespace FluentNHibernate.Utils
 {
-    internal static class HbmExtensions
+    public static class HbmExtensions
     {
         public static HbmType ToHbmType(this TypeReference typeReference)
         {
             return new HbmType()
             {
                 name = typeReference.Name,
+            };
+        }
+
+        public static HbmFilterParam ToHbmFilterParam(this KeyValuePair<string, IType> parameterPair)
+        {
+            return new HbmFilterParam()
+            {
+                name = parameterPair.Key,
+                type = parameterPair.Value.Name
             };
         }
     }
